Avoid printf on records
`strings.Join()` uses `strings.Builder{}` for performance the same way;
thus I'll go ahead and claim this is more performant.

BTW the Go std lib special cases the string "+"-operator when the number
of concatenating operands is 5.  In this case it is 6.  Thus I assume
there is no more performant way of concatenating this except twiddling
with unsafe directly.