clusters_tests: lower connection timeout
Host in the test are invalid anyway. And NewClusters time out for that
period of time trying to connect to them.
I'm lowering the timeout so we don't have this test run for 1 minute.
It might be better if we use hosts from the 192.0.2.0/24  class in tests
in the future.